Update Airbrake 4.3 -> 7.4
It was not possible to update Airbrake while keeping its configuration
backwards-compatible, as the updated version requires the 'project_id'
parameter to be defined. As the changes are not backwards-compatible
anyways, it was decided to also 'modernize' all other parameters
so they would match Airbrake configuration parameter names.

- `project_id` was added (required)
- `api_key` was renamed to `project_key` (required)
- `environment_name` was renamed to `environment` (required)
- `ignore_environments` was added (optional)
  - Airbrake does not filter by `test and `development` environments
    implicitly anymore, so be sure to use this parameter if you
    don't want to send error reports from such environments.
    Accepts strings, symbols and regexes.
- `host` is same as before (required)
- `notifier_name` was added (optional)
  - useful when multiple Airbrake notifiers are used in one project

Airbrake client does not use the `async` flag anymore, so it was dropped
from the factory, too (`Airbrake.notify` already notifies asynchronously).
Same with the `secure` flag - sending the reports is secure by default.